requirements.txt
Flask>=1.0.2
Flask-Cors>=3.0.3
mecab-python3>=0.7
gunicorn==20.1.0
